Revert "Mule 13492 (#5695)"
This reverts commit ddac66d281e26c5eddba8d90960d1ae84f12e43f.

+++ core/src/test/java/org/mule/routing/CollectionMessageSplitterTestCase.java
@@ -7,6 +7,7 @@
 package org.mule.routing;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertSame;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
@@ -200,7 +201,7 @@ public class CollectionMessageSplitterTestCase extends AbstractMuleContextTestCa
         {
             assertTrue(msg.getPayload() instanceof String);
             assertEquals(counted ? count : -1,
-                         (int)msg.getOutboundProperty(MuleProperties.MULE_CORRELATION_GROUP_SIZE_PROPERTY));
+                msg.getOutboundProperty(MuleProperties.MULE_CORRELATION_GROUP_SIZE_PROPERTY));
             actualSequences.add(msg.getOutboundProperty(MuleProperties.MULE_CORRELATION_SEQUENCE_PROPERTY));
             String str = (String)msg.getPayload();
             assertTrue(TEST_LIST_MULTIPLE.contains(str));
